[fix] Rename cmd modules to avoid conflict with built-in cmd
Rename the module defined in 'codechecker_common/cmd' to
'codechecker_common/cli_commands' (as there is already a cli.py file
in the same folder) in order to avoid ambiguity with Python's
built-in cmd module when using `pdb.set_trace()`. The new name
better reflects its purpose and eliminates potential naming
conflicts.

Rename all 'cmd' directories to 'cli' for consistency with
'codechecker_common/cli_commands'.
