+++ analyzer/codechecker_analyzer/cli/analyze.py
