+++ analyzer/codechecker_analyzer/cli/analyzer_version.py
