+++ analyzer/codechecker_analyzer/cli/checkers.py
