+++ codechecker_common/cli_commands/version.py
@@ -67,7 +67,7 @@ def main(args):
     # Get analyzer version information if the module is available.
     analyzer_version = None
     try:
-        from codechecker_analyzer.cmd.analyzer_version import Version
+        from codechecker_analyzer.cli.analyzer_version import Version
         analyzer_version = Version()
     except Exception:
         pass
@@ -75,7 +75,7 @@ def main(args):
     # Get web version information if the module is available.
     web_version = None
     try:
-        from codechecker_web.cmd.web_version import Version
+        from codechecker_web.cli.web_version import Version
         web_version = Version()
     except Exception:
         pass
