+++ web/client/codechecker_client/cli/cmd.py
