+++ web/client/codechecker_client/cli/store.py
